compute_metrics.py: added some time tracking stuff graph_2d.py: removed some redundant code graph_3d.py: removed some redundant code jitter_plots.py: some new datasets and plots metrics_gd.py: added deviation metrics to crossing res and angular res, adjusted max number of crossings possible, adjusted node node occlusion metric to work with shapely, adjusted occlusion metrics to their original (measures total area of intersection vs possible intersection areas) parallel_bar_plot.py: removed some redundant code projections.py: added a line to not recompute existing projections visualization_tool.py: added more comments

+++ projections.py
@@ -3,6 +3,7 @@ import pandas as pd
 import time
 import constants
 import warnings
+import numpy as np
 
 from sklearn.manifold import TSNE
 
@@ -15,18 +16,19 @@ techniques = set(data_frame['layout_technique'].to_list())
 consolid_metrics_file = os.path.join(constants.metrics_dir, 'metrics.pkl')
 
 for ds in datasets:
-    print('Starting projections for the metrics of ' + str(ds))
-    projs_list = []
-    start = time.time()
-    for tech in techniques:
-        curr_metrics = data_frame[(data_frame['dataset_name'] == ds) & (data_frame['layout_technique'] == tech) & (data_frame['n_components'] == 3)]
-        metrics_2d = curr_metrics[constants.metrics].to_numpy()[0]
-        # add metrics of 2d
-        project = TSNE(n_components = 2, perplexity = 100).fit_transform(np.vstack([curr_metrics['views_metrics'].to_numpy()[0], metrics_2d]))
-        projs_list.append([ds, tech, project])
+    if not os.path.isfile(constants.metrics_projects_dir + '/' + ds.split('.')[0] + '_projcs.pkl'):
+        print('Starting projections for the metrics of ' + str(ds))
+        projs_list = []
+        start = time.time()
+        for tech in techniques:
+            curr_metrics = data_frame[(data_frame['dataset_name'] == ds) & (data_frame['layout_technique'] == tech) & (data_frame['n_components'] == 3)]
+            metrics_2d = curr_metrics[constants.metrics].to_numpy()[0]
+            # add metrics of 2d
+            project = TSNE(n_components = 2, perplexity = 100).fit_transform(np.vstack([curr_metrics['views_metrics'].to_numpy()[0], metrics_2d]))
+            projs_list.append([ds, tech, project])
 
-    print('Creating projections took: ' + str(round(time.time() - start, 2)) + 's')
+        print('Creating projections took: ' + str(round(time.time() - start, 2)) + 's')
 
-    df_projs = pd.DataFrame.from_records(projs_list)
-    df_projs.columns = ['dataset_name', 'layout_technique', 'projection']
-    df_projs.to_pickle(constants.metrics_projects_dir + '/' + ds.split('.')[0] + '_projcs.pkl')
+        df_projs = pd.DataFrame.from_records(projs_list)
+        df_projs.columns = ['dataset_name', 'layout_technique', 'projection']
+        df_projs.to_pickle(constants.metrics_projects_dir + '/' + ds.split('.')[0] + '_projcs.pkl')
